Move service file into services dir
