+++ apps/api/src/app/modules/mail/mail.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { MailService } from './mail.service';
+import { MailService } from './services/mail.service';
 
 
 @Module({
