+++ apps/api/src/app/modules/playlist/controllers/playlist-controller/playlist.controller.spec.ts
@@ -3,7 +3,7 @@ import { PlaylistController } from './playlist.controller';
 import { PlaylistService } from '../../services/playlist/playlist.service';
 import { SpotifyService } from '../../../spotify/spotify.service';
 import { PlaylistHistoryService } from '../../services/playlist-history/playlist-history.service';
-import { MailService } from '../../../mail/mail.service';
+import { MailService } from '../../../mail/services/mail.service';
 
 describe('PlaylistController', () => {
   let controller: PlaylistController;
